Emulated Wiimote: Updated the accelerometer recording to record the gravity forces in all directions instead of recording raw data, to bypass problem of different Wiimotes having different neutral values
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2316 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_Wiimote/Src/main.h
@@ -51,9 +51,9 @@ void OpenConsole(bool Open = false);
 #define WM_RECORDING_IR 2
 struct SRecording
 {
-	u8 x;
-	u8 y;
-	u8 z;
+	int x;
+	int y;
+	int z;
 	double Time;
 	u8 IR[12];
 };
